+++ node/impl/eth/api.go
@@ -228,6 +228,7 @@ type MpoolAPI interface {
 // MessagePool is a minimal version of messagepool.MessagePool
 type MessagePool interface {
 	PendingFor(ctx context.Context, a address.Address) ([]*types.SignedMessage, *types.TipSet)
+	Pending(ctx context.Context) ([]*types.SignedMessage, *types.TipSet)
 	GetConfig() *types.MpoolConfig
 }